RAG_assisted_OCR/requirements.txt
paddlepaddle
paddleocr
numpy
matplotlib
opencv-python
langchain_core 
langchain_community 
uvicorn 
pypdf 
langchain-huggingface 
faiss-cpu 
langchainhub
sentence_transformers 
PyPDF2
fastapi
streamlit
